+++ openneuro/__init__.py
@@ -4,6 +4,10 @@ except ImportError:
     # Running on pre-3.8 Python; use importlib-metadata package
     import importlib_metadata as metadata
 
-__version__ = metadata.version('openneuro-py')
+try:
+    __version__ = metadata.version("openneuro-py")
+except metadata.PackageNotFoundError:
+    # package is not installed
+    pass
 
 from .download import download  # noqa: F401